fix: self-invocation 이슈를 해결
@Transactional이 있는 메소드를 CheckInTransactionService 컴포를 만들어서 분리했어요. 디렉토리를 어디에 해야 할 지 몰라서 걍 냅뒀는데, 옮길 필요가 있다면 옮겨주시기 바랍니다..
