+++ src/main/kotlin/rainbowfriends/daramserverv2/domain/checkin/component/CheckInStatusSwitch.kt
@@ -1,20 +1,14 @@
 package rainbowfriends.daramserverv2.domain.checkin.component
 
 import org.springframework.stereotype.Component
-import org.springframework.transaction.annotation.Transactional
-import rainbowfriends.daramserverv2.domain.member.exception.MemberNotFoundException
-import rainbowfriends.daramserverv2.global.checkin.repository.CheckInRepository
-import rainbowfriends.daramserverv2.global.checkin.entity.CheckIn
-import rainbowfriends.daramserverv2.global.member.entity.Member
-import rainbowfriends.daramserverv2.global.member.repository.MemberRepository
-import rainbowfriends.daramserverv2.global.redis.redisson.util.DistributedLock
+import rainbowfriends.daramserverv2.domain.checkin.service.CheckInTransactionService
 import rainbowfriends.daramserverv2.global.checkin.component.CheckInDataSync
+import rainbowfriends.daramserverv2.global.redis.redisson.util.DistributedLock
 import java.time.LocalDate
 
 @Component
 class CheckInStatusSwitch(
-    private val checkInRepository: CheckInRepository,
-    private val memberRepository: MemberRepository,
+    private val checkInTransactionService: CheckInTransactionService,
     private val checkInDataSync: CheckInDataSync,
     private val distributedLock: DistributedLock
 ) {
@@ -29,36 +23,13 @@ class CheckInStatusSwitch(
 
     private fun handleCheckInStatusSwitch(studentId: Short, date: LocalDate): Boolean {
         return try {
-            val user = getMemberInfo(studentId)
-            val checkIn = getCheckInRecord(user, date)
-            toggleCheckInStatus(checkIn)
+            val user = checkInTransactionService.getMemberInfo(studentId)
+            val checkIn = checkInTransactionService.getCheckInRecord(user, date)
+            checkInTransactionService.toggleCheckInStatus(checkIn)
+            checkInDataSync.syncCheckInToMongoDB(checkIn)
             true
         } catch (e: Exception) {
             false
         }
-    }
-
-    @Transactional
-    fun getMemberInfo(studentId: Short): Member {
-        studentId.toString().let {
-            val grade = it[0].toString().toInt()
-            val classNum = it[1].toString().toInt()
-            val studentNum = it.substring(2).toInt()
-            return memberRepository.findByGradeAndClassNumAndNumber(grade, classNum, studentNum)
-                ?: throw MemberNotFoundException("Member not found")
-        }
-    }
-
-    @Transactional
-    fun getCheckInRecord(user: Member, date: LocalDate): CheckIn {
-        return checkInRepository.findByUserAndCheckinDate(user, date)
-            ?: throw IllegalStateException("Check-in record not found")
-    }
-
-    @Transactional
-    fun toggleCheckInStatus(checkIn: CheckIn) {
-        checkIn.checkinStatus = !checkIn.checkinStatus
-        checkInRepository.save(checkIn)
-        checkInDataSync.syncCheckInToMongoDB(checkIn)
     }
 }
